movies_admin/requirements/base.txt
django==3.1
django-model-utils==4.0.0
python-decouple==3.3
psycopg2==2.8.6
gunicorn==20.0.4